+++ components/common/Footer/index.tsx
